NBT

Documentation for NBT.

NBT.Tag

NBT.get_tags

NBT.get_tags

NBT.read_nbt_uncompressed

NBT.set_tags

NBT.set_tags

NBT.write_nbt_uncompressed

NBT.Tag — Type

struct Tag{T}

Represents an NBT tag, containing an id, a name, and data.

The following are all values of T, ordered by id, starting from 0x1, and their corresponding names on the wiki:

[UInt8, Int16, Int32, Int64, Float32, Float64, UInt8[], String, Tag[], Tag[], Int32[], Int64[] ]

[Byte, Short, Int, Long, Float, Double, Byte_Array, String, List, Compound, Int_Array, Long_Array]

Properties

id::UInt8: The id of the tag. See minecraft wiki.

name::String: The name of the tag.

data::T: The data in the tag. Its type is dictated by id. Do not use the wrong id-T combination or everything will break!

NBT.get_tags — Method

get_tags(tag, id::Integer; depth=10)

Returns a Vector{Tag} containing all Tags in tag with id id, with an optional search depth limit.

For convenience, getindex is implemented, and only gets the first match on depth 1:

julia> tag = read("xd.litematic", Tag);

julia> tag[0x3] === get_tags(tag, 0x3; depth=1)[1]
true

See also get_tags

NBT.get_tags — Method

get_tags(tag, name::String; depth=10)

Returns a Vector{Tag} containing all Tags in tag named name, with an optional search depth limit.

For convenience, getindex is implemented, and only gets the first match on depth 1:

julia> pwd()

julia> tag = read("xd.litematic", Tag);

julia> tag["Version"] === get_tags(tag, "Version"; depth=1)[1]
true

See also get_tags

NBT.read_nbt_uncompressed — Method

read_nbt_uncompressed(io, ::Type{Tag})

Reads an nbt tag from an uncompressed IO. Not exported.

NBT.set_tags — Method

set_tags(tag, id::Integer, newtag; depth=10)

Sets all Tags in tag with id id to newtag, with an optional search depth limit.

For convenience, setindex! is implemented, and only sets the first match on depth 1:

julia> tag = read("xd.litematic", Tag);

julia> tag[0x3] = Tag(0x3, "EEEE", 69420)
Int32 EEEE: 69420

See also get_tags

NBT.set_tags — Method

set_tags(tag, name::String, newtag; depth=10)

Sets all Tags in tag named name to newtag, with an optional search depth limit.

For convenience, setindex! is implemented, and only sets the first match on depth 1:

julia> tag = read("xd.litematic", Tag);

julia> tag["Version"] = Tag(0x3, "Version", 69420)
Int32 Version: 69420

See also get_tags

NBT.write_nbt_uncompressed — Method

write_nbt_uncompressed(io, tag)

Writes an nbt tag to an uncompressed IO. Not exported.